Migrating to SYSFW version w2020.17-am64x
Only am64x related binaries are updated. Other SoCs are not tested, so
those binaries remain unchanged from v2020.08b release.

+++ packages/ti/drv/sciclient/soc/sysfw/include/am65x/tisci_resasg_types.h
@@ -74,8 +74,8 @@
  */
 
 #define TISCI_RESASG_UTYPE(type, subtype) \
-    (((type << TISCI_RESASG_TYPE_SHIFT) & TISCI_RESASG_TYPE_MASK) | \
-     ((subtype << TISCI_RESASG_SUBTYPE_SHIFT) & TISCI_RESASG_SUBTYPE_MASK))
+    ((((uint16_t) (type) << TISCI_RESASG_TYPE_SHIFT) & TISCI_RESASG_TYPE_MASK) |    \
+     (((uint16_t) (subtype) << TISCI_RESASG_SUBTYPE_SHIFT) & TISCI_RESASG_SUBTYPE_MASK))
 
 /**
  * IA subtypes definitions
